Admin Dashboard/requirements.txt
streamlit>=1.32.0
spacy>=3.7.0
sentence-transformers>=2.6.0
networkx>=3.2.1
matplotlib>=3.8.0
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.2.0
plotly>=5.20.0
pyvis>=0.3.2
torch>=2.1.0
transformers>=4.38.0
PyJWT>=2.8.0
bcrypt>=4.1.0
wikipedia>=1.4.0
arxiv>=2.1.0
